Rename and remove commented code.
